went through, looking for small errors here and there, found a few, sorta fixed em

+++ src/addressingModes.c
@@ -3,7 +3,7 @@
 
 // The addressing modes do not offset the PC, this is handled by the CPU
 
-int ABS_A(int PC) { return readByte(PC) + readByte(PC + 1) << 8; }
+int ABS_A(int PC) { return readByte(PC) + ((int)readByte(PC + 1) << 8); }
 int ABS_INDEX_X(int PC) {return PC + readByte(getCPU_XRegister());}
 int ABS_INDEX_Y(int PC) {return PC + readByte(getCPU_YRegister());}
 int ACC(int PC) {return getCPU_Accumulator();}
@@ -14,7 +14,7 @@ int ZP(int PC)  {return readByte(PC);}
 int ZP_INDX_IND(int PC) {
     int indirectAddr = readByte(PC) + readByte(getCPU_XRegister());
     indirectAddr &= 0xFF;
-    return readByte(indirectAddr) + readByte(indirectAddr + 1) << 8;
+    return readByte(indirectAddr) + ((int)readByte(indirectAddr + 1) << 8);
 }
 int ZP_INDX_X(int PC) {
     int effectiveAddr = readByte(PC) + readByte(getCPU_XRegister());
@@ -26,13 +26,13 @@ int ZP_INDX_Y(int PC){
 }
 int ZP_IND(int PC) {
     int indirectAddr = readByte(PC);
-    return readByte(indirectAddr) + readByte(indirectAddr + 1) << 8;
+    return readByte(indirectAddr) + ((int)readByte(indirectAddr + 1) << 8);
 }
 
 int ZP_IND_INDX_Y(int PC) {
     int indirectAddr = readByte(PC) + readByte(getCPU_YRegister());
     indirectAddr &= 0xFF;
-    return readByte(indirectAddr) + readByte(indirectAddr + 1) << 8;
+    return readByte(indirectAddr) + ((int)readByte(indirectAddr + 1) << 8);
 }
 
 /*
